+++ src/domain/authentication/exceptions/authentication.exception.ts
