+++ src/domain/items/exceptions/items.exception.ts
@@ -1,6 +1,6 @@
 import { BaseException } from '@/lib/exceptions/base.exception';
 import { HttpStatus } from '@nestjs/common';
-import { ItemExceptionEnum } from '../exception.enum';
+import { ItemExceptionEnum } from '../../../lib/exceptions/exception.enum';
 
 export class ItemNotFoundException extends BaseException {
   constructor(message: string) {
